Split component in Entity and Entry List
